feat: move globals to seperate folder
