+++ src/theme/globals/colors.ts
@@ -1,0 +1,6 @@
+import { paletteConfig } from 'theme/palette.config';
+import { enhancePaletteWithShades } from 'theme/utils';
+
+const colors = enhancePaletteWithShades(paletteConfig);
+
+export default colors;
