+++ src/theme/globals/constants/textCase.ts
