+++ src/theme/globals/index.ts
@@ -1,0 +1,23 @@
+import borderRadius from './borderRadius';
+import borderWidth from './borderWidth';
+import boxShadow from './boxShadow';
+import colors from './colors';
+import elevation from './elevation';
+import opacity from './opacity';
+import sizing from './sizing';
+import spacing from './spacing';
+import typography from './typography';
+
+const globals = {
+  colors,
+  typography,
+  spacing,
+  elevation,
+  borderRadius,
+  borderWidth,
+  boxShadow,
+  opacity,
+  sizing,
+};
+
+export default globals;
