+++ src/theme/index.ts
@@ -1,38 +1,17 @@
-import borderRadius from './borderRadius';
-import borderWidth from './borderWidth';
-import boxShadow from './boxShadow';
-import elevation from './elevation';
-import opacity from './opacity';
+import globals from './globals';
 import overrides from './overrides';
 import { getAAColor, getAAColorFromSwatches, getColor } from './palette';
-import { paletteConfig } from './palette.config';
-import sizing from './sizing';
-import spacing from './spacing';
 import { ColorScheme, TextColorTypes, Theme, ThemeConfig } from './types';
-import typography from './typography';
-import { enhancePaletteWithShades } from './utils';
 
 const defaultTheme = (theming: ColorScheme): Theme => {
-  const palette = enhancePaletteWithShades(paletteConfig);
-
   return {
-    globals: {
-      colors: palette,
-      typography,
-      spacing,
-      elevation,
-      borderRadius,
-      borderWidth,
-      boxShadow,
-      opacity,
-      sizing,
-    },
+    globals,
     colorScheme: theming,
     overrides,
     utils: {
-      getColor: getColor(palette),
-      getAAColorFromSwatches: getAAColorFromSwatches(palette),
-      getAAColor: getAAColor(palette),
+      getColor: getColor(globals.colors),
+      getAAColorFromSwatches: getAAColorFromSwatches(globals.colors),
+      getAAColor: getAAColor(globals.colors),
     },
   };
 };
